+++ src/main.ts
@@ -5,5 +5,15 @@ export type {
   PluginHttpRequest,
   PluginHttpResponse,
   PluginIpc,
+  PluginScriptCollectionInit,
+  PluginScriptContext,
+  PluginScriptContextInit,
+  PluginScriptEnvironmentInit,
+  PluginScriptKeyValue,
+  PluginScriptRequestInit,
+  PluginScriptResponseInit,
+  PluginScriptRunResult,
+  PluginScripts,
+  PluginScriptTestResult,
   PluginStorage
 } from './types';